+++ R/venndir-legend.R
@@ -1,13 +1,11 @@
 
-#' Add optional legend to venndir figures (developmental)
+#' Add legend to venndir figure
 #' 
-#' Add optional legend to venndir figures (developmental). This function
-#' is being evaluated and developed, and will very likely change
-#' again in an upcoming release.
+#' Add legend to venndir figure
 #' 
 #' Note this function is experimental and is under active development. The
 #' implementation and arguments may change in future.
-#' 
+#'  
 #' Limitations: Currently this function relies upon having the `setlist`
 #' used to produce the `venndir()` output, and the `venndir()` output.
 #' In future, the `setlist` could be derived from the `venndir()` output
@@ -28,9 +26,15 @@
 #' 
 #' @family venndir utility
 #' 
-#' @param venndir_output `Venndir` object returned by `venndir()`,
-#'    which is used to generate the legend counts.
-#'    When supplied, the `set_colors` are also defined by this object.
+#' @param venndir_output `Venndir` object, default NULL, or one of:
+#'    * `Venndir` object, also used to define `set_colors` when applicable
+#'    * `list` representing `setlist`
+#'    * `data.frame` output from `signed_overlaps()`
+#' @param setlist `list`, default NULL, used when `venndir_output` is NULL
+#'    * `list` representing `setlist`
+#'    * `data.frame` with output from `signed_overlaps()`, with columns
+#'    `"sets"`, `"overlap"`, `"count"`. It may include `"each"` which is
+#'    required to display signed counts with `legend_signed=TRUE`.
 #' @param x `character` string indicating the position of the legend,
 #'    as passed to `graphics::legend()` when `legend_style="base"`.
 #'    Not relevant when `legend_style="data.frame"`.
@@ -38,11 +42,24 @@
 #'    should match `names(setlist)`. When not supplied, colors are inferred
 #'    from `venndir_output`, and when that is not supplied, colors are
 #'    defined using `colorjam::rainbowJam()`.
+#' @param legend_style `character` string, default `"grid"` with the
+#'    legend style, with these options:
+#'    * `"grid"` generates a grid `grob` using `gridExtra::grid.table()`.
+#'    * `"data.frame"` generates the legend `data.frame`.
 #' @param unicode `logical` default TRUE, whether to use unicode arrows
 #'    for signed counts, used when `legend_signed=TRUE`.
 #'    This argument is passed to `curate_venn_labels()`.
-#' @param setlist `list` alternative when `venndir_output` is not supplied
-#'    as a `Venndir` object. Used internally by `textvenn()`.
+#' @param legend_headers `character` vector with headers to use,
+#'    named by the default heading. To display no column header,
+#'    set `legend_headers=NULL` or `legend_headers=FALSE`.
+#'    Recognized names:
+#'    * "Set" - the header for the set names, default "Set".
+#'    * "Label" - when `labels` or `legend_labels` are defined, the
+#'    header "Label" can be customized.
+#'    * "Size" - the header for the set sizes, default "Size".
+#'    * "Sign" - the header for the set signs, default "Sign".
+#'    * "Percentage" - the header for the set size percentage overall,
+#'    default "Percentage".
 #' @param keep_newlines `logical` indicating whether to keep newlines
 #'    (line breaks) in the set labels used in the Venn legend.
 #' @param legend_total `logical` default TRUE, whether to include the
@@ -76,9 +93,6 @@
 #'    * `"blackborder"`: will use `"#000000"` border color, matching the Total.
 #' @param box.lwd `numeric` used to define the box line width,
 #'    as passed to `graphics::legend()` when `legend_style="base"`.
-#' @param legend_style `character` string indicating the style of legend:
-#'    * `"grid"` uses `gridExtra::grid.table()`
-#'    * `"data.frame"` only returns the venndir legend `data.frame`.
 #' @param item_type `character` string used as a suffix to the counts
 #'    displayed in the legend. Use `item_type=""` or `item_type=NULL`
 #'    to hide this suffix.
@@ -91,33 +105,36 @@
 #'    * `header_border` defines the border color for the header, which
 #'    is also applied to all cells in the table. Note that the line is
 #'    drawn with line width `lwd`.
-#' @param lwd `numeric` value for line width used for the cell border
+#' @param lwd `numeric`, default 1, cell border line width used
 #'    in the legend table, when `legend_style="grid"`.
-#' @param x_inset,y_inset `grid::unit` object that defines the inset
-#'    distance away from each plot edge, along the x-axis and y-axis,
+#' @param x_inset,y_inset `grid::unit`, default 2 lines each, used
+#'    when `legend_style="grid"` to position the legend relative to `x`,
+#'    useful when positioning the legend near the edge `x="bottomleft"`.
+#'    The inset controls distance from the plot edge, along the x- and y-axes,
 #'    respectively. For example `x_inset=grid::unit(2, "lines")` will
 #'    place the legend table 2 character lines (which are defined by
 #'    line height for typical character size) away from the left or
-#'    right edge of the plot. Any valid `grid::unit` can be used, however
-#'    using `"lines"` units provides some consistent width rather
-#'    than using `"npc"` in terms of fraction of overall plot width
-#'    or height.
+#'    right edge of the plot. Any valid `grid::unit` can be used.
+#'    Note that the metric is "plot edge", and the plot is fixed 1:1 aspect,
+#'    so for wide plot devices a negative `x_inset` may be used to nudge
+#'    the legend outside the typical plot boundary.
 #' @param font_cex `numeric` adjustment to default font sizes. The default
 #'    font size with `legend_style="grid"` uses a 12 point font, so to adjust
 #'    to a specific font size like 8 points, use: `font_cex=8/12`
 #' @param fontfamily `character` used as the default font for all labels,
 #'    and is used in `fontfamilies` by default. However `fontfamilies`
 #'    takes priority when defined.
-#' @param fontfamilies `list` with named elements: `"overlap"`, `"count"`,
-#'    `"signed"`, to allow a custom font for each element of the legend.
-#'    This argument is also used by `assemble_venndir_label()`.
-#'    * The "overlap" font is used for set names, and column headers.
-#'    Column headers are currently always bold.
-#'    * Note that "overlap" is used for `"Size"` and `"Percent"` columns, and
-#'    when `combine_size=TRUE` the signed counts are in the same cell
-#'    and therefore use the same font.
-#'    * When `combine_size=FALSE` the signed counts use "signed".
-#' @param table_theme `list` of theme parameters as described in
+#' @param fontfamilies `list` with named elements
+#'    to allow a custom font for each element of the legend, consistent
+#'    with usage by `assemble_venndir_label()`.
+#'    * `"overlap"`: used for set labels and column headers,
+#'    and column headers are currently always bold font.
+#'    * `"count"`: used for main counts in "Size", and percentage in
+#'    "Percentage" when displayed.
+#'    * `"signed"`: used for signed counts, when shown and when
+#'    `combine_size=FALSE`, otherwise the signed labels are combined
+#'    with count labels and they re-use the same font as `"count"` above.
+#' @param table_theme `list` default NULL, optional theme as described in
 #'    `gridExtra::tableGrob()`, and `gridExtra::ttheme_default()`.
 #'    When supplied, the `font_cex` argument is ignored. 
 #'    The `list` components include:
@@ -127,23 +144,24 @@
 #'    * `parse` - `logical` whether to parse plotmath expressions
 #'    * `padding` - `grid::unit()` for horizontal and vertical padding
 #'    within each cell
-#' @param draw_legend `logical` indicating whether to draw the resulting
-#'    legend. Because this function is in development, the `legend_style="grid"`
-#'    seems to be preferred, however it may be useful to return the
-#'    `grob` object for manipulation before rendering inside the active
-#'    plot. In that case, use arguments: `legend_style="grid", draw_legend=FALSE`.
-#' @param alias `character` vector with optional set aliases, experimental
-#'    feature to allow referring to entries in `setlist` by shorter aliases.
-#'    Note that `names(alias)` must also match `names(setlist)`.
-#'    Note: The argument `labels` is more likely to be useful, since
-#'    the original call to `venndir()` should use `setlist` with aliased
-#'    names, and expanded labels would be provided separately.
-#' @param labels `character` vector with optional set labels, experimental
-#'    feature to supply a longer label associated with each entry in `setlist`.
-#'    Note that `names(labels)` must also match `names(setlist)`.
-#'    Note: The most useful workflow would be to assign short aliases to
-#'    `names(setlist)` when calling `venndir()`, then use the original long
-#'    label names as argument `labels` here.
+#' @param draw_legend `logical`, default TRUE, whether to draw the resulting
+#'    legend. When `FALSE` the data is returned based upon `legend_style`,
+#'    when `legend_style="grid"` it returns a grid `grob`, otherwise
+#'    it returns `data.frame`.
+#' @param alias `character`, default NULL, optional set aliases,
+#'    intended to add a new first column to the legend with these
+#'    aliases. Useful to display a short label in the legend table,
+#'    while also supplying a longer label via `names(setlist)`
+#'    or `labels`.
+#'    This argument differs from `labels` in that `labels` replaces
+#'    the `names(setlist)` in the legend table with `labels`.
+#'    However, `alias` adds a new column to the table as column 1.
+#' @param labels `character` default NULL, optional custom labels to
+#'    display in place of `names(setlist)` in the legend table.
+#'    Note: `names(labels)` must also match `names(setlist)`.
+#'    This argument differs from `labels` in that `labels` replaces
+#'    the `names(setlist)` in the legend table with `labels`.
+#'    However, `alias` adds a new column to the table as column 1.
 #' @param legend_padding `numeric` or `grid::unit` used to define padding for
 #'    each table cell. This value is only used when `table_theme` is not
 #'    provided.
@@ -167,6 +185,12 @@
 #' # move to different corner
 #' vo <- venndir(setlist, legend_x="bottomleft")
 #' 
+#' # add alias column
+#' vo <- venndir(setlist,
+#'    legend_labels=setNames(paste0("Items detected in ", LETTERS[1:3]),
+#'       names(setlist)),
+#'    alias=setNames(LETTERS[1:3], names(setlist)))
+#' 
 #' # turn off the default legend
 #' vo <- venndir(setlist, draw_legend=FALSE)
 #' venndir_legender(setlist=setlist,
@@ -176,20 +200,54 @@
 #' # test multi-line labels
 #' names(setlist) <- c("Group B-<br>Group A",
 #'    "Group C-<br>\nGroup B",
-#'    "Group_C-<br>\nGroupA")
+#'    "Group C-<br>\nGroup A")
 #' vo <- venndir(setlist,
 #'    draw_legend=FALSE,
+#'    font_cex=1.3,
+#'    fontfamily="Arial",
+#'    expand_fraction=c(-0.1, 0.4, -0.1, -0.1),
 #'    show_segments=FALSE)
 #' venndir_legender(setlist=setlist,
 #'    venndir_output=vo,
+#'    font_cex=1.3,
 #'    x="bottomleft")
-#' 
-#' # Same as above, showing how to render the legend_grob.
-#' # This method also "hides" the column headers.
+#'    
+#' # some common customizations
 #' vo <- venndir(setlist,
-#'    show_segments=FALSE,
 #'    draw_legend=FALSE,
-#'    font_cex=c(1, 1, 0.5, 0.5))
+#'    expand_fraction=c(0.2, 0.2, 0.2, 0.2)*3,
+#'    font_cex=0.7,
+#'    fontfamily="Arial")
+#' venndir_legender(vo,
+#'    legend_headers=c(Set="Comparison",
+#'       Size="Counts (Signed Counts)"),
+#'    legend_color_style=c("nofill", "blackborder"),
+#'    x="bottomleft")
+#' venndir_legender(vo,
+#'    combine_size=FALSE,
+#'    legend_headers=c(Set="Comparison",
+#'       Size="Counts", Sign="(Signed Counts)"),
+#'    legend_color_style=c("nofill", "blackborder"),
+#'    x="bottomright")
+#' venndir_legender(vo,
+#'    combine_size=FALSE,
+#'    combine_sign=FALSE,
+#'    legend_headers=c(Set="Comparison",
+#'       Size="Counts", Sign="(Signed Counts)"),
+#'    legend_color_style=c("nofill", "blackborder"),
+#'    x="topright")
+#' venndir_legender(vo,
+#'    legend_percent=TRUE,
+#'    legend_headers=c(Set="Comparison",
+#'    Percentage="Pct.",
+#'       Size="Counts", Sign="(Signed Counts)"),
+#'    legend_color_style=c("nofill", "blackborder"),
+#'    x="topleft")
+#' 
+#' # Example showing how to render the legend_grob
+#' # which may offer some flexibility.
+#' vo <- venndir(setlist,
+#'    draw_legend=FALSE)
 #' legend_grob <- venndir_legender(setlist=setlist,
 #'    venndir_output=vo,
 #'    draw_legend=FALSE,
@@ -232,10 +290,16 @@
 #' @export
 venndir_legender <- function
 (venndir_output=NULL, 
- x="bottomleft", 
  setlist=NULL,
+ x="bottomleft", 
  set_colors=NULL,
+ legend_style=c("grid",
+    "data.frame"),
  unicode=TRUE,
+ legend_headers=c(Set="Set",
+    Size="Size",
+    Percentage="Percentage",
+    Sign="Sign"),
  sign_count_delim=NULL,
  keep_newlines=FALSE,
  legend_total=TRUE,
@@ -245,9 +309,7 @@ venndir_legender <- function
  combine_size=TRUE,
  legend_color_style=c("fill",
     "border"),
- box.lwd=0, 
- legend_style=c("grid",
-    "data.frame"),
+ box.lwd=0,
  item_type="",
  header_color="#000000",
  header_bg="#FFFFFF00",
@@ -273,15 +335,115 @@ venndir_legender <- function
 {
    legend_style <- match.arg(legend_style);
 
+   legend_headers_default <- eval(formals(venndir_legender)$legend_headers);
+   if (length(legend_headers) == 0 || FALSE %in% legend_headers) {
+      legend_headers <- legend_headers_default;
+      legend_headers[] <- "";
+   }
+   if (!all(names(legend_headers_default) %in% names(legend_headers))) {
+      add_names <- setdiff(names(legend_headers_default),
+         names(legend_headers));
+      legend_headers[add_names] <- legend_headers_default[add_names];
+   }
+   
+   # handle input data
    if (length(venndir_output) > 0) {
-      if (inherits(venndir_output, "Venndir")) {
+      # recognize first argument venndir_output as list or data.frame
+      # by assigning to setlist
+      if (inherits(venndir_output, c("list", "data.frame"))) {
+         if (length(setlist) == 0) {
+            setlist <- venndir_output;
+            venndir_output <- NULL;
+         } else {
+            # consider warning that venndir_output is ignored
+            if (verbose) {
+               # Todo: use cli::
+               jamba::printDebug("venndir_legender(): ",
+                  "venndir_output is ", c("list", "data.frame"),
+                  " and ", "setlist", " is provided, so ",
+                  "venndir_output", " is ignored.")
+            }
+            venndir_output <- NULL;
+         }
+      } else if (inherits(venndir_output, "Venndir")) {
+         if (length(setlist) > 0) {
+            # consider warning that it will be overwritten
+         }
          setlist <- venndir_output@setlist;
       } else {
-         stop("venndir_output must have class 'Venndir'.")
+         stop(paste0("When supplied, venndir_output must be ",
+            "'Venndir', or 'list', or 'data.frame'"));
       }
    } else if (length(setlist) == 0) {
-      stop("Either setlist or venndir_output must be supplied.")
+      stop("Either venndir_output or setlist are required as input.")
    }
+   
+   # handle setlist as signed_overlaps() output, convert to setlist
+   if (inherits(setlist, "data.frame")) {
+      if (!all(c("sets", "overlap", "count") %in% colnames(setlist))) {
+         stop(paste0("Input as a data.frame must contain colnames: ",
+            jamba::cPaste(c("'sets'", "'overlap'", "'count'")),
+            ", optionally 'each'."));
+      }
+      so <- setlist;
+      if (all(c("each", "overlap_label") %in% colnames(so))) {
+         # for 'each' we can infer all signed overlaps
+         # otherwise we must handle it as unsigned
+         sol <- lapply(split(so, so$sets), function(idf){
+            jamba::nameVector(idf$count,
+               gsub(" ", "_", idf$overlap_label))
+         })
+         # convert to setlist
+         setlist <- signed_counts2setlist(sol)
+         # define default legend_signed
+         if (length(legend_signed) == 0) {
+            legend_signed <- TRUE;
+         }
+      } else {
+         # without 'each' column, we must handle it as unsigned counts
+         soc <- lapply(split(so, so$sets), function(idf){
+            sum(idf$count)
+         })
+         setlist <- counts2setlist(soc)
+         # define default legend_signed
+         if (length(legend_signed) == 0) {
+            legend_signed <- FALSE;
+         }
+      }
+   }
+   
+   # detect default legend_signed
+   if (length(legend_signed) == 0) {
+      if (length(venndir_output) > 0) {
+         # by default, follow what is shown in the venndir_output
+         if ("overlap" %in% metadata(venndir_output)$overlap_type) {
+            legend_signed <- FALSE;
+         } else {
+            legend_signed <- ("signed" %in% venndir_output@label_df$type);
+         }
+      } else {
+         # observe setlist, look for signed data with multiple unique values
+         is_signed <- any(unlist(lapply(setlist, function(i){
+            if (length(names(i)) == 0) {
+               FALSE
+            } else {
+               length(setdiff(i, c("", "0", NA))) > 1
+            }
+         })))
+         if (TRUE %in% is_signed) {
+            legend_signed <- TRUE;
+         }
+      }
+   }
+   # turn off legend_signed when there are no signed counts?
+   # for now, skip this logic
+   # if (length(venndir_output) > 0 &&
+   #       TRUE %in% legend_signed &&
+   #       !"signed" %in% venndir_output@label_df$type) {
+   #    legend_signed <- FALSE
+   # }
+   
+   # define empty sign_count_delim
    if (length(sign_count_delim) == 0) {
       if (length(venndir_output) > 0) {
          if ("sign_count_delim" %in% names(metadata(venndir_output))) {
@@ -292,26 +454,7 @@ venndir_legender <- function
          sign_count_delim <- " ";
       }
    }
-   
-   # detect legend_signed
-   if (length(legend_signed) == 0) {
-      if (length(venndir_output) > 0) {
-         if ("overlap" %in% metadata(venndir_output)$overlap_type) {
-            legend_signed <- FALSE
-         } else if (length(metadata(venndir_output)$overlap_type) == 1 &&
-               "signed" %in% venndir_output@label_df$type) {
-            legend_signed <- TRUE;
-         } else {
-            legend_signed <- FALSE;
-         }
-      }
-   }
-   # turn off legend_signed when there are no signed counts
-   if (TRUE %in% legend_signed && length(venndir_output) > 0 &&
-         !"signed" %in% venndir_output@label_df$type) {
-      legend_signed <- FALSE
-   }
-   
+
    # detect legend_percentage
    if (length(legend_percentage) == 0) {
       legend_percentage <- FALSE;
@@ -323,24 +466,29 @@ venndir_legender <- function
    }
    
    # make sure setlist matches Venn displayed sets
-   legend_labels <- NULL;
-   if (length(venndir_output) > 0) {
-      use_setlist_names <- intersect(names(setlist),
-         venndir_output@label_df$overlap_set)
-      if (length(use_setlist_names) == 0) {
-         stop("label_df$overlap_set does not match names(setlist)");
-      }
-      setlist <- setlist[use_setlist_names];
-      if ("legend_label" %in% colnames(venndir_output@jps@polygons)) {
-         matchset <- match(names(setlist),
-            venndir_output@jps@polygons$venn_name)
-         if (any(is.na(matchset))) {
-            stop("Venndir 'venn_name' does not match names(setlist).");
-            # Todo: recover by ignoring legend_labels
-         } else {
-            legend_labels <- jamba::nameVector(
-               venndir_output@jps@polygons$legend_label[matchset],
-               names(setlist));
+   if (length(labels) > 0) {
+      legend_labels <- labels;
+      labels <- NULL;
+   } else {
+      legend_labels <- NULL;
+      if (length(venndir_output) > 0) {
+         use_setlist_names <- intersect(names(setlist),
+            venndir_output@label_df$overlap_set)
+         if (length(use_setlist_names) == 0) {
+            stop("label_df$overlap_set does not match names(setlist)");
+         }
+         setlist <- setlist[use_setlist_names];
+         if ("legend_label" %in% colnames(venndir_output@jps@polygons)) {
+            matchset <- match(names(setlist),
+               venndir_output@jps@polygons$venn_name)
+            if (any(is.na(matchset))) {
+               stop("Venndir 'venn_name' does not match names(setlist).");
+               # Todo: recover by ignoring legend_labels
+            } else {
+               legend_labels <- jamba::nameVector(
+                  venndir_output@jps@polygons$legend_label[matchset],
+                  names(setlist));
+            }
          }
       }
    }
@@ -376,19 +524,37 @@ venndir_legender <- function
          stop("Either setlist or venndir_output must be provided.");
       }
       setlist <- setlist(venndir_output);
-      ## old logic
-      # vodf <- venndir_output@jps@polygons;
-      # vodf1 <- subset(vodf, type %in% "overlap");
-      # total_items <- sum(vodf1$venn_counts);
-      # # vodf1[, c("venn_name", "venn_counts")]
-      # vodf_names <- strsplit(vodf1$venn_name, "&");
-      # vodf_levels <- unique(unlist(vodf_names));
-      # vodf_counts <- rep(vodf1$venn_counts, lengths(vodf_names));
-      # vodf_split <- factor(unlist(vodf_names), levels=vodf_levels);
-      # setlist_sizes <- sapply(split(vodf_counts, vodf_split), sum);
-      # setlist <- lapply(setlist_sizes, function(i){
-      #    rep(1, length.out=i)
-      # })
+   }
+   
+   # Quick function to update legend colnames
+   adjust_legend_headers <- function
+   (gridlegend_df_use,
+    legend_headers,
+    cell_padding=c("", ""),
+    ...)
+   {
+      # repair slight hack with trailing whitespace
+      # used for column justification
+      colnames(gridlegend_df_use) <- gsub("^Size ", "Size",
+         colnames(gridlegend_df_use));
+      # rename only when legend_headers contains something to change
+      if (!all(names(legend_headers) == legend_headers)) {
+         switch_from <- intersect(colnames(gridlegend_df_use),
+            names(legend_headers))
+         if (length(switch_from) > 0) {
+            switch_to <- legend_headers[switch_from];
+            gridlegend_df_use <- jamba::renameColumn(gridlegend_df_use,
+               from=switch_from,
+               to=switch_to,
+               check.names=FALSE)
+         }
+      }
+      # add padding
+      colnames(gridlegend_df_use) <- paste0(
+         cell_padding[[1]],
+         colnames(gridlegend_df_use),
+         cell_padding[[2]])
+      gridlegend_df_use
    }
    
    # pick out venn colors if available
@@ -468,7 +634,7 @@ venndir_legender <- function
          # venndir_legender(setlist=setlist, venndir_output=vo, x="bottomright")
          total_items <- sum(vodf_all$venn_counts, na.rm=TRUE);
       } else {
-         sv <- signed_overlaps(setlist[sets],
+         sv <- signed_overlaps(setlist,
             overlap_type="overlap",
             return_items=FALSE,
             ...);
@@ -490,6 +656,11 @@ venndir_legender <- function
    setlist <- fix_setlist_names(setlist,
       use_newline="\n",
       keep_newlines=keep_newlines)
+   if (length(alias) > 0) {
+      alias <- fix_setlist_names(alias,
+         use_newline="\n",
+         keep_newlines=keep_newlines)
+   }
    vodf_color <- fix_setlist_names(vodf_color,
       use_newline="\n",
       keep_newlines=keep_newlines)
@@ -609,6 +780,7 @@ venndir_legender <- function
          set=names(setlist),
          label=labels[names(setlist)],
          size=lengths(setlist))
+      rownames(gridlegend_df) <- names(setlist)
    }
    
    # optional alias
@@ -622,11 +794,15 @@ venndir_legender <- function
          jamba::formatInt(lengths(setlist)),
          " ",
          item_type)
-      gridlegend_df <- data.frame(check.names=FALSE,
-         stringsAsFactors=FALSE,
-         set=alias[names(setlist)],
-         label=names(setlist),
-         size=lengths(setlist))
+      gridlegend_df$label <- gridlegend_df$set;
+      gridlegend_df$set <- alias[names(setlist)];
+      gridlegend_df <- gridlegend_df[, c("set", "label", "size"), drop=FALSE];
+      rownames(gridlegend_df) <- names(setlist)
+      # gridlegend_df <- data.frame(check.names=FALSE,
+      #    stringsAsFactors=FALSE,
+      #    set=alias[names(setlist)],
+      #    label=names(setlist),
+      #    size=lengths(setlist))
    }
    
    # optionally include total item count
@@ -640,7 +816,11 @@ venndir_legender <- function
       if (!"label" %in% colnames(gridlegend_df)) {
          gridlegend_df2$label <- NULL;
       }
+      rownames(gridlegend_df2) <- "Total";
       gridlegend_df <- rbind(gridlegend_df, gridlegend_df2);
+      # confirm no duplicate rownames, and do no rename first instance
+      rownames(gridlegend_df) <- jamba::makeNames(
+         rownames(gridlegend_df), renameFirst=FALSE);
       #
       legend_df2 <- data.frame(legend="Total",
          # color="grey95",
@@ -675,7 +855,8 @@ venndir_legender <- function
          1000 * gridlegend_df$size / tail(gridlegend_df$size, 1)) / 10;
       gcdf2p <- data.frame(check.names=FALSE,
          `Percentage`=paste(format(pct_total, digits=3, trim=TRUE), "%"))
-      rownames(gcdf2p) <- gridlegend_df$set;
+      # rownames(gcdf2p) <- gridlegend_df$set;
+      rownames(gcdf2p) <- rownames(gridlegend_df);
       if (TRUE %in% legend_signed) {
          matchgc <- match(rownames(gcdf2p), rownames(gcdf2))
          gcdf2 <- data.frame(check.names=FALSE,
@@ -690,9 +871,9 @@ venndir_legender <- function
    # add signed/percentage columns to legend
    if (TRUE %in% legend_signed || TRUE %in% legend_percentage) {
       matchgc <- match(gridlegend_df$set, rownames(gcdf2));
-      add_df <- gcdf2[match(gridlegend_df$set, rownames(gcdf2)), , drop=FALSE];
+      add_df <- gcdf2[match(rownames(gridlegend_df), rownames(gcdf2)), , drop=FALSE];
       add_df[is.na(add_df)] <- ""
-      addc_df <- gcdf1[match(gridlegend_df$set, rownames(gcdf2)), , drop=FALSE];
+      addc_df <- gcdf1[match(rownames(gridlegend_df), rownames(gcdf2)), , drop=FALSE];
       addc_df[is.na(addc_df)] <- ""
       gridlegend_df[, colnames(add_df)] <- add_df
    }
@@ -737,6 +918,29 @@ venndir_legender <- function
             colnames(gridlegend_df_use));
          rownames(gridlegend_df_use) <- seq_len(
             nrow(gridlegend_df_use));
+         # Bonus points: Apply column justification
+         use_justify <- ifelse(
+            grepl("Size$|Percent", colnames(gridlegend_df_use)),
+            "right", "left")
+         justify_colnames <- TRUE;
+         # Adjust colnames by legend_headers
+         gridlegend_df_use <- adjust_legend_headers(gridlegend_df_use,
+            legend_headers=legend_headers,
+            cell_padding=c("", ""))
+         # Bonus points: Apply column justification
+         for (inum in seq_len(ncol(gridlegend_df_use))) {
+            if (TRUE %in% justify_colnames) {
+               usex <- c(colnames(gridlegend_df_use)[inum],
+                  gridlegend_df_use[[inum]])
+               gridlegend_df_use[[inum]] <- tail(format(usex,
+                  justify=use_justify[[inum]]), -1);
+               colnames(gridlegend_df_use)[inum] <- head(format(usex,
+                  justify=use_justify[[inum]]), 1);
+            } else {
+               gridlegend_df_use[[inum]] <- format(gridlegend_df_use[[inum]],
+                  justify=use_justify[[inum]])
+            }
+         }
          return(gridlegend_df_use)
       }
       
@@ -749,22 +953,14 @@ venndir_legender <- function
       #   be placed at the border of the extended padded range, the very edge.
       colnames(gridlegend_df_use) <- jamba::ucfirst(
          colnames(gridlegend_df_use));
-      # for (icol in head(colnames(gridlegend_df_use), -1)) {
-      # add leading whitespace to all columns
+      
+      # add leading,trailing whitespace to all columns
+      cell_padding <- c("  ", " ");
       for (icol in colnames(gridlegend_df_use)) {
-         gridlegend_df_use[[icol]] <- paste0("  ", gridlegend_df_use[[icol]]);
-         inum <- match(icol, colnames(gridlegend_df_use))
-         colnames(gridlegend_df_use)[inum] <- paste0("  ",
-            colnames(gridlegend_df_use)[inum]);
-      }
-      # for (icol in tail(colnames(gridlegend_df_use), 1)) {
-      # add trailing whitespace to all columns
-      for (icol in colnames(gridlegend_df_use)) {
-         gridlegend_df_use[[icol]] <- paste0(gridlegend_df_use[[icol]], " ");
-         inum <- match(icol, colnames(gridlegend_df_use))
-         colnames(gridlegend_df_use)[inum] <- paste0(
-            colnames(gridlegend_df_use)[inum],
-            " ");
+         # data.frame cells
+         gridlegend_df_use[[icol]] <- paste0(cell_padding[[1]],
+            gridlegend_df_use[[icol]],
+            cell_padding[[2]]);
       }
       if (length(table_theme) == 0) {
          use_hjust1 <- ifelse(
@@ -782,6 +978,7 @@ venndir_legender <- function
             legend_padding <- rep(legend_padding, length.out=2);
             use_padding <- grid::unit(legend_padding, "mm");
          }
+         
          table_theme <- gridExtra::ttheme_default(
             base_size=12 * font_cex,
             base_family=fontfamilies$overlap,
@@ -799,6 +996,11 @@ venndir_legender <- function
             # colhead=list(fg_params=list(hjust=1, x=0.95)))
       }
       
+      # Final updates to colnames
+      gridlegend_df_use <- adjust_legend_headers(gridlegend_df_use,
+         legend_headers=legend_headers,
+         cell_padding=cell_padding)
+
       # create the tableGrob
       legend_grob <- gridExtra::tableGrob(
          rows=NULL,